requirements.txt
flask==2.2.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pymongo==4.3.3
flask-cors==3.0.10
python-dotenv==1.0.0
langchain==0.0.267
langchain_nvidia_ai_endpoints
langchain_core
pydantic==1.10.8
langchain_community
python-json-logger==2.0.7
langraph==0.0.27
typing-extensions==4.7.1